Improve add reference workflow, move component to the new place.
