+++ src/article/editor/styles/_index.css
@@ -18,7 +18,6 @@
 @import './_math.css';
 @import './_fn-group.css';
 @import './_ref-list.css';
-@import './_add-reference.css';
 @import './_xref';
 @import './_fig.css';
 @import './_insert-fig-tool.css';
