+++ src/article/metadata/ReferenceUploadComponent.js
