+++ src/article/metadata/styles/_add-reference-workflow.css
@@ -1,7 +1,3 @@
-.se-add-reference {
-
-}
-
 .se-add-reference > .se-title {
   font-size: 20px;
   padding: 20px;
@@ -19,8 +15,11 @@
 }
 
 .se-add-reference > .se-doi-input > .se-input-group > .se-input {
+  width: 100%;
   background: #f1f1f1;
   padding: 7px 80px 7px 10px;
+  border: none;
+  border-radius: 4px;
 }
 
 .se-add-reference > .se-doi-input > .se-input-group > .se-action {
@@ -145,6 +144,7 @@
   border: 1px solid #e1e1e1;
   border-radius: 4px;
   text-align: center;
+  line-height: 30px;
   cursor: pointer;
 }
