+++ src/article/metadata/styles/_index.css
@@ -7,5 +7,6 @@
 @import '/_metadata-editor.css';
 @import '/_entity-editor.css';
 @import '/_translatable-editor.css';
+@import '/_add-reference-workflow.css';
 @import '/_form-row.css';
 @import '/_card.css';
